+++ host/sgx/quote.c
@@ -247,13 +247,13 @@ oe_result_t sgx_verify_quote(
     if (p_qve_report_info != NULL)
     {
         OE_TRACE_INFO(
-            "SGX DCAP QvE-based quote verification is used, res: %s\n",
+            "SGX DCAP QvE-based SGX quote verification is used, res: %s\n",
             oe_result_str(result));
     }
     else
     {
         OE_TRACE_INFO(
-            "SGX DCAP QVL-based quote verification is used, res: %s\n",
+            "SGX DCAP QVL-based SGX quote verification is used, res: %s\n",
             oe_result_str(result));
     }
 done: